+++ utils/collection/queue/queue.go
@@ -3,6 +3,8 @@ package queue
 import (
 	"iter"
 	"slices"
+
+	"github.com/ARM-software/golang-utils/utils/collection"
 )
 
 // NewQueue returns a Queue which is not thread safe
@@ -78,7 +80,7 @@ func (s *Queue[T]) Enqueue(value ...T) {
 }
 
 func (s *Queue[T]) EnqueueSequence(seq iter.Seq[T]) {
-	for v := range seq {
+	for v := range collection.SequenceOrEmpty(seq) {
 		s.enqueue(v)
 	}
 }