+++ utils/collection/stack/stack.go
@@ -3,6 +3,8 @@ package stack
 import (
 	"iter"
 	"slices"
+
+	"github.com/ARM-software/golang-utils/utils/collection"
 )
 
 // NewStack returns a stack which is not thread safe
@@ -73,7 +75,7 @@ func (s *Stack[T]) Push(value ...T) {
 }
 
 func (s *Stack[T]) PushSequence(seq iter.Seq[T]) {
-	for v := range seq {
+	for v := range collection.SequenceOrEmpty(seq) {
 		s.push(v)
 	}
 }